Handle private images.
