+++ pkg/controller/stack/utils/stacks.go
@@ -2,11 +2,12 @@ package utils
 
 import (
 	"fmt"
+	"net/url"
+	"regexp"
 	"strings"
 
-	"github.com/google/go-containerregistry/pkg/name"
-	"github.com/google/go-containerregistry/pkg/v1/remote"
 	kabanerov1alpha2 "github.com/kabanero-io/kabanero-operator/pkg/apis/kabanero/v1alpha2"
+	corev1 "k8s.io/api/core/v1"
 )
 
 // Removes the tag portion of all images associated with the input stack version.
@@ -22,58 +23,56 @@ func RemoveTagFromStackImages(stack *kabanerov1alpha2.StackVersion, stackName st
 	return nil
 }
 
-// Retrieves the repository part of input image that contains both a repo path and tag.
-// An error is returned if the image input is empty.
-func GetImageRepository(image string) (string, error) {
-	if len(image) == 0 {
-		return "", fmt.Errorf("The input image is empty.")
+// Custom filter method that allows the retrieval of a secret containing an annotation with a key of the
+// form [string]-[number] (i.e kabanero.io/git-0), and a value that contains a hostname/domain (i.e. github.com).
+// If there are multiple matching secrets, the annotation with the lexically lowest value key is used.
+// If no secret matches annotation key, but there are secrets with an annotation value
+// that matches the hostname/domain, the first one seen is used. If a secret could not be found, nil is returned.
+func SecretAnnotationFilter(secretList *corev1.SecretList, filterStrings ...string) (*corev1.Secret, error) {
+	var keyMatchingSecret *corev1.Secret = nil
+	var noKeyMatchingSecret *corev1.Secret = nil
+	kabKey := ""
+	for _, secret := range secretList.Items {
+		annotations := secret.GetAnnotations()
+		for key, value := range annotations {
+			matchedUrl, err := regexp.MatchString("^https?://", value)
+			if err != nil {
+				return nil, err
+			}
+			if matchedUrl {
+				surl, err := url.Parse(value)
+				if err != nil {
+					fmt.Println("Unable to parse secret annotation value URL: ", surl, ". Secret: ", secret)
+				}
+
+				if surl.Hostname() == filterStrings[0] {
+					if strings.HasPrefix(key, filterStrings[1]) {
+						if len(kabKey) == 0 {
+							kabKey = key
+							keyMatchingSecret = secret.DeepCopy()
+						} else {
+							if kabKey > key {
+								kabKey = key
+								keyMatchingSecret = secret.DeepCopy()
+							}
+						}
+					} else {
+						// Save the first secret we see matching the hostname.
+						if noKeyMatchingSecret == nil {
+							noKeyMatchingSecret = secret.DeepCopy()
+						}
+					}
+				}
+			}
+		}
 	}
 
-	// A tag name must be valid ASCII and may contain lowercase and uppercase
-	// letters, digits, underscores, periods and dashes.  A tag name may not
-	// start with a period or a dash, and may contain a maximum of 128
-	// characters.
-	//
-	// The tag cannot contain a slash.  So, if we find a slash past the last
-	// colon, what we found is not a tag and we should leave it alone.
-	//
-	// Additionally, the components of a repository name cannot contain
-	// a colon, with the exception of the first component which is actually
-	// a hostname.  So if the repository contains a colon which is not the
-	// tag separator, it must occur in the first component, and there must
-	// also be a second component after the first (implying there is a
-	// slash between the components).
-	//
-	// Reference (docker-specific, but should apply to others):
-	//   https://github.com/docker/distribution/blob/release/2.7/reference/reference.go
-	//   https://docs.docker.com/engine/reference/commandline/tag/
-
-	repo := image
-	tagIndex := strings.LastIndex(image, ":")
-	slashIndex := strings.LastIndex(image, "/")
-	if (tagIndex >= 0) && (slashIndex < tagIndex) {
-		repo = image[0:tagIndex]
+	if keyMatchingSecret != nil {
+		return keyMatchingSecret, nil
+	}
+	if noKeyMatchingSecret != nil {
+		return noKeyMatchingSecret, nil
 	}
 
-	return repo, nil
-}
-
-// Retrieves the input image digest from the hosting repository.
-func RetrieveImageDigest(image string) (string, error) {
-	ref, err := name.ParseReference(image, name.WeakValidation)
-	if err != nil {
-		return "", err
-	}
-
-	img, err := remote.Image(ref)
-	if err != nil {
-		return "", err
-	}
-
-	h, err := img.Digest()
-	if err != nil {
-		return "", err
-	}
-
-	return h.Hex, nil
+	return nil, nil
 }
